"댓글"

+++ src/main/java/company/friendsdog/dogcommunity/entity/Reply.java
@@ -11,10 +11,12 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+//펫번호,내용,댓글번호
+// int 말고 Long
 public class Reply {
-    private int boardNo;
-    private int petNo;
-    private int replyNo;
+    private Long boardNo;
+    private Long petNo;
+    private Long replyNo;
     private String comment;
     private LocalDateTime replyDateTime;
 }
